Split up local/remote steps
